Apply refactoring to API.
